refactor: Change Main to Duke and Duke to TaskList
Some scripts are dependent on the main entrypoint being Duke
